chore: migrate ERC20 token creation to web3-shared-evm
