+++ packages/mask/src/plugins/Trader/constants/quickswap.ts
@@ -1,5 +1,4 @@
-import { ChainId } from '@masknet/web3-shared-evm'
-import { DAI, ETHER, maUSDC, QUICK, USDC, USDT, WBTC, WNATIVE, WNATIVE_ONLY } from './trader'
+import { ChainId, DAI, ETHER, maUSDC, QUICK, USDC, USDT, WBTC, WNATIVE, WNATIVE_ONLY } from '@masknet/web3-shared-evm'
 import type { ERC20AgainstToken, ERC20TokenCustomizedBase } from './types'
 
 /**
