+++ packages/mask/src/plugins/Trader/constants/sashimiswap.ts
@@ -1,5 +1,4 @@
-import { ChainId } from '@masknet/web3-shared-evm'
-import { COMP, DAI, MKR, USDC, USDT, WNATIVE, WNATIVE_ONLY } from './trader'
+import { ChainId, COMP, DAI, MKR, USDC, USDT, WNATIVE, WNATIVE_ONLY } from '@masknet/web3-shared-evm'
 import type { ERC20AgainstToken, ERC20TokenCustomizedBase } from './types'
 
 /**
