+++ packages/mask/src/plugins/Trader/constants/uniswap.ts
@@ -1,7 +1,6 @@
-import { ChainId } from '@masknet/web3-shared-evm'
+import { ChainId, AMPL, DAI, MSKA, MSKB, MSKC, USDC, USDT, WBTC, WNATIVE, WNATIVE_ONLY } from '@masknet/web3-shared-evm'
 import { Percent } from '@uniswap/sdk-core'
 import JSBI from 'jsbi'
-import { AMPL, DAI, MSKA, MSKB, MSKC, USDC, USDT, WBTC, WNATIVE, WNATIVE_ONLY } from './trader'
 import type { ERC20AgainstToken, ERC20TokenCustomizedBase } from './types'
 
 /**
