+++ packages/mask/src/plugins/Trader/constants/wannaswap.ts
@@ -1,5 +1,4 @@
-import { ChainId } from '@masknet/web3-shared-evm'
-import { DAI, USDC, USDT, WBTC, WANNA, WNATIVE, WNATIVE_ONLY } from './trader'
+import { ChainId, DAI, USDC, USDT, WBTC, WANNA, WNATIVE, WNATIVE_ONLY } from '@masknet/web3-shared-evm'
 import type { ERC20AgainstToken, ERC20TokenCustomizedBase } from './types'
 
 /**
