+++ packages/web3-shared/evm/constants/index.ts
@@ -1,2 +1,3 @@
 export * from './primitives'
 export * from './constants'
+export * from './tokens'
